Fixed a bug in the version number
Changed version number to 1.0.1 instead of 1.0.3.

Version number will be influenced by relevance of changes. For example, a core change or complete rewrite will result in the first number changing, a major change but smaller than a rewrite will result in the second changing, and a minor change for the third.